moved input components into own folder
