+++ frontend/src/components/formInputs/GithubInput/GithubInput.js
@@ -1,0 +1,8 @@
+import React from "react"
+import BasicInput from "../BasicInput/BasicInput"
+
+const GithubInput = props => (
+  <BasicInput componentName="github" label="github" type="text" {...props} />
+)
+
+export default GithubInput
