+++ frontend/src/components/formInputs/TagsInput/TagsInput.js
@@ -1,0 +1,8 @@
+import React from "react"
+import BasicInput from "../BasicInput/BasicInput"
+
+const TagsInput = props => (
+  <BasicInput componentName="tags" label="tags" type="text" {...props} />
+)
+
+export default TagsInput
